add projectRoot and rename packageManagerControllerBase file
