+++ packages/cli-core/src/package-manager-controller/npm_package_manager_controller.ts
@@ -3,7 +3,7 @@ import _fsp from 'fs/promises';
 import { execa as _execa } from 'execa';
 import * as _path from 'path';
 import { executeWithDebugLogger as _executeWithDebugLogger } from './execute_with_logger.js';
-import { PackageManagerControllerBase } from './package_manager_controller.js';
+import { PackageManagerControllerBase } from './package_manager_controller_base.js';
 
 /**
  * NpmPackageManagerController is an abstraction around npm commands that are needed to initialize a project and install dependencies
