+++ packages/cli-core/src/package-manager-controller/package_manager_controller_base.ts
