+++ packages/cli-core/src/package-manager-controller/pnpm_package_manager_controller.ts
@@ -3,7 +3,7 @@ import * as _path from 'path';
 import { existsSync as _existsSync } from 'fs';
 import { execa as _execa } from 'execa';
 import { executeWithDebugLogger as _executeWithDebugLogger } from './execute_with_logger.js';
-import { PackageManagerControllerBase } from './package_manager_controller.js';
+import { PackageManagerControllerBase } from './package_manager_controller_base.js';
 
 /**
  * PnpmPackageManagerController is an abstraction around pnpm commands that are needed to initialize a project and install dependencies
