+++ packages/create-amplify/src/amplify_project_creator.ts
@@ -1,4 +1,5 @@
 import { LogLevel } from '@aws-amplify/cli-core';
+import { PackageManagerController } from '@aws-amplify/plugin-types';
 import { ProjectRootValidator } from './project_root_validator.js';
 import { GitIgnoreInitializer } from './gitignore_initializer.js';
 import { InitialProjectFileGenerator } from './initial_project_file_generator.js';
@@ -26,7 +27,7 @@ export class AmplifyProjectCreator {
    * Delegates out to other classes that handle parts of the getting started experience
    */
   constructor(
-    private readonly packageManagerController: any,
+    private readonly packageManagerController: PackageManagerController,
     private readonly projectRootValidator: ProjectRootValidator,
     private readonly gitIgnoreInitializer: GitIgnoreInitializer,
     private readonly initialProjectFileGenerator: InitialProjectFileGenerator
@@ -71,19 +72,14 @@ export class AmplifyProjectCreator {
       process.cwd() === this.packageManagerController.projectRoot
         ? ''
         : `Navigate to your project directory using
-'cd .${
-            this.packageManagerController.projectRoot.replace(
-              process.cwd(),
-              ''
-            ) as string
-          }'.
+'cd .${this.packageManagerController.projectRoot.replace(process.cwd(), '')}'.
 Then get started with the following commands:
 `;
 
     printer.log(
       `Welcome to AWS Amplify! 
 ${cdPreamble}
-${this.packageManagerController.getWelcomeMessage() as string}`
+${this.packageManagerController.getWelcomeMessage()}`
     );
 
     printer.log(
