+++ packages/plugin-types/API.md
@@ -141,6 +141,7 @@ export type MainStackNameResolver = {
 
 // @public (undocumented)
 export type PackageManagerController = {
+    projectRoot: string;
     getWelcomeMessage: () => string;
     initializeProject: () => Promise<void>;
     initializeTsConfig: (targetDir: string) => Promise<void>;
